feat: relative path bonus
* feat: add relative path to rm, cp, mv

* chore: fix memory leak (unfinished)

* chore: fix memory leak

* feat: relative path to exec

* chore: fix exec filename

+++ src/user/syscall.c
@@ -11,22 +11,22 @@ void syscall(uint32_t eax, uint32_t ebx, uint32_t ecx, uint32_t edx)
     __asm__ volatile("int $0x30");
 }
 
-void read_file_api(struct FAT32DriverRequest *req, uint32_t *code)
+void read_file_api(struct FAT32DriverRequest *req, int8_t *code)
 {
     syscall(0, (uint32_t)req, (uint32_t)code, 0);
 }
 
-void read_directory_api(struct FAT32DriverRequest *req, uint32_t *code)
+void read_directory_api(struct FAT32DriverRequest *req, int8_t *code)
 {
     syscall(1, (uint32_t)req, (uint32_t)code, 0);
 }
 
-void write_api(struct FAT32DriverRequest *req, uint32_t *code)
+void write_api(struct FAT32DriverRequest *req, int8_t *code)
 {
     syscall(2, (uint32_t)req, (uint32_t)code, 0);
 }
 
-void delete_api(struct FAT32DriverRequest *req, uint32_t *code)
+void delete_api(struct FAT32DriverRequest *req, int8_t *code)
 {
     syscall(3, (uint32_t)req, (uint32_t)code, 0);
 }
@@ -66,7 +66,7 @@ void set_cursor_col(uint32_t col)
     syscall(10, col, 0, 0);
 }
 
-void search_file_api(struct FAT32DriverRequest *req, uint32_t *code)
+void search_file_api(struct FAT32DriverRequest *req, int8_t *code)
 {
     syscall(11, (uint32_t)req, (uint32_t)code, 0);
 }
@@ -76,17 +76,17 @@ void get_cluster_number_api(struct FAT32DriverRequest *req, uint32_t *code)
     syscall(12, (uint32_t)req, (uint32_t)code, 0);
 }
 
-void copy_folder_api(uint32_t src, uint32_t dest, uint32_t *code)
+void copy_folder_api(uint32_t src, uint32_t dest, int8_t *code)
 {
     syscall(13, src, dest, (uint32_t)code);
 }
 
-void copy_file_api(struct FAT32DriverRequest *src, struct FAT32DriverRequest *dest, uint32_t *code)
+void copy_file_api(struct FAT32DriverRequest *src, struct FAT32DriverRequest *dest, int8_t *code)
 {
     syscall(14, (uint32_t)src, (uint32_t)dest, (uint32_t)code);
 }
 
-void delete_recursive_api(struct FAT32DriverRequest *req, uint32_t *code)
+void delete_recursive_api(struct FAT32DriverRequest *req, int8_t *code)
 {
     syscall(15, (uint32_t)req, (uint32_t)code, 0);
 }
@@ -132,7 +132,7 @@ void put_char_position(uint32_t row, uint32_t col, char c)
     syscall(24, (uint32_t)row, (uint32_t)col, (uint32_t)c);
 }
 
-void get_time_stamp(uint32_t * time)
+void get_time_stamp(uint32_t *time)
 {
     syscall(25, (uint32_t)time, 0, 0);
 }
